+++ CHANGES.rst
@@ -3,6 +3,7 @@
 - [FIXED] Fixed ``TypeError`` when setting revision limits on Python>=3.6.
 - [FIXED] Fixed the ``exists()`` double check on ``client.py`` and ``database.py``.
 - [FIXED] Fixed Cloudant exception code 409 with 412 when creating a database that already exists.
+- [NEW] Added support for ``_local`` documents by adding the ``LocalDocument`` class.
 
 2.4.0 (2017-02-14)
 ==================